Appify

The Appify module builds a shiny app that displays dynamic graphical and tabular representations of the input data (occurrence and covarite) as well as model predictions.

Appify(.model, .ras, directory = tempdir(), run_App = FALSE)

Arguments

.model

Internal parameter, do not use in the workflow function. .model is list of a data frame (data) and a model object (model). .model is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.#'

.ras

Internal parameter, do not use in the workflow function. .ras is a raster layer, brick or stack object. .ras is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.#'

directory

(Optional) a string giving the path to a directory within which to save the shiny app. This directory must exist.#'

run_App

Logical, should the app be run automatically. Not recommended for workflows containing lists. Default is FALSE

Details

Appify will return the path to the shiny application. For details on how to share the application visit http://shiny.rstudio.com/tutorial/lesson7/. You can run the app locally using shiny::runApp(appDir = <YOUR_PATH>)

Version

1.0

Date submitted

2016-05-05

Data type

presence-only, presence/absence, presence/background, abundance, proportion

See also

Other output: InteractiveCovariateMap, InteractiveMap, NoOutput, PerformanceMeasures, PredictNewRasterMap, PrintMap, ROCcurve, ResponseCurveViz, ResponseCurve, SameTimePlaceMap, SeparatePA, SurfaceMap, VariableImportance

Author
ZOON Developers, Tom August, zoonproject@gmail.com