Process module: Background

Process module to generate random (optionally biased) background records in cells of the covariate raster and return these along with the presence only data.

Background(.data, n = 100, bias = NULL, seed = NULL)

Arguments

.data

Internal parameter, do not use in the workflow function. .data is a list of a data frame and a raster object returned from occurrence modules and covariate modules respectively. .data is passed automatically in workflow from the occurrence and covariate modules to the process module(s) and should not be passed by the user.

n

the number of background points to sample

bias

optional RasterLayer with cells giving the relative probability of a background record being sampled there. Alternatively, a length one numeric giving a radius (in KM) around presence points to take background points from. If bias = NULL (the default) then no biasing is applied, and all non-missing cells are equally likely to be selected.

seed

Numeric used with set.seed

Version

1.1

Date submitted

2016-03-23

Data type

presence-only

See also

Other process: AddRandomUniformPredictors, BackgroundAndCrossvalid, Bootstrap, Clean, Crossvalidate, JitterOccurrence, MESSMask, NoProcess, OneHundredBackground, OneThousandBackground, PartitionDisc, StandardiseCov, TargetGroupBackground, Transform, addInteraction

Author
ZOON Developers, zoonproject@gmail.com