Covariate module: Bioclim

Get worldclim environment data. Downloads then stores locally.

Bioclim(extent = c(-180, 180, -90, 90), resolution = 10, layers = 1:5)

Arguments

extent

Length 4 numeric vector giving min longitude, max longitude, min latitude, max latitude.

resolution

Resolution in minutes. Must be one of 2.5, 5 or 10. Default is 10.

layers

which bioclim layers to obtain, a vector of integers between 1 and 19.

Version

1.0

Date submitted

2015-11-13

See also

Other covariate: AirNCEP, Bioclim_future, CarolinaWrenRasters, LocalRaster, NCEP, NaiveRandomRaster, UKAir, UKBioclim

Author
ZOON Developers, zoonproject@gmail.com