Covariate module: Bioclim_future

Get worldclim data for the future (CMIP5). Downloads then stores locally.

Bioclim_future(extent = c(-10, 10, 45, 65), resolution = 10,
  layers = 1:19, rcp = 45, model = "AC", year = 70)

Arguments

extent

Length 4 numeric vector giving min longitude, max longitude, min latitude, max latitude.

resolution

Resolution in minutes. Must be one of 2.5, 5 or 10. Default is 10.

layers

which bioclim layers to obtain, a vector of integers between 1 and 19. See http://www.worldclim.org/bioclim.

rcp

numeric. Representative Concentration Pathways, or greenhouse gases concentration trajectories (see https://en.wikipedia.org/wiki/Representative_Concentration_Pathways. Should be one of 26, 45, 60, or 85.

model

character. General Circulation Model. Should be one of "AC", "BC", "CC", "CE", "CN", "GF", "GD", "GS", "HD", "HG", "HE", "IN", "IP", "MI", "MR", "MC", "MP", "MG", or "NO".

year

numeric. Time period. Should be either 50 (which represents the period 2041-2060) or 70 (which represents the period 2061-2080).

Version

1.0.0

Date submitted

2016-06-15

See also

http://www.worldclim.org and getData.

Other covariate: AirNCEP, Bioclim, CarolinaWrenRasters, LocalRaster, NCEP, NaiveRandomRaster, UKAir, UKBioclim

Author
F. Rodriguez-Sanchez & E. Van Loon, zoonproject@gmail.com