Presence/absence data for the Carolina Wren in the USA

Presence/absence of the Carolina Wren Thryothorus ludovicianus at Breeding Bird Survey transects in 2006. These data are made available in the maxlike R package. See carw.data for more details. This module loads them data transform the coordinates to lat/longs and formats the data for use in a zoon workflow.

CarolinaWrenPA()

Details

Data type

presence/absence

Version

0

Date submitted

2016-05-21

See also

Other occurrence: AnophelesPlumbeus, CWBZimbabwe, CarolinaWrenPO, LocalOccurrenceData, Lorem_ipsum_UK, NATrees, NBNdataByName, NaiveRandomPresenceAbsence, NaiveRandomPresence, SpOcc, SugarMaple, UKAnophelesPlumbeus

Author
Nick Golding, nick.golding.research@gmail.com