Process module: Crossvalidate

Run k fold crossvalidation. If presence absence, split presences and absences separately so folds have equally balanced data. Otherwise just sample.

Crossvalidate(.data, k = 5, seed = NULL)

Arguments

.data

Internal parameter, do not use in the workflow function. .data is a list of a data frame and a raster object returned from occurrence modules and covariate modules respectively. .data is passed automatically in workflow from the occurrence and covariate modules to the process module(s) and should not be passed by the user.

k

Positive integer number of folds to split the data into. Default is 5.

seed

Numeric used with set.seed

Version

1.0

Date submitted

2015-11-13

Data type

presence-only, presence/absence, abundance, proportion

See also

Other process: AddRandomUniformPredictors, BackgroundAndCrossvalid, Background, Bootstrap, Clean, JitterOccurrence, MESSMask, NoProcess, OneHundredBackground, OneThousandBackground, PartitionDisc, StandardiseCov, TargetGroupBackground, Transform, addInteraction

Author
ZOON Developers, zoonproject@gmail.com