Add noise to occurrence coordinates

Often the geographic coordinates of occurrence records are subject to uncertainty. One approach ot accounting for that uncertainty in an SDM analysis is to run several analyses, each time sampling a different location from a probability distribution over the likely "true" coordinates. JitterOccurrence generates one such sample, under an assumption that the distribution is an isotropic Gaussian with standard deviation sd

JitterOccurrence(.data, sd = 0)

Arguments

.data

Internal parameter, do not use in the workflow function. .data is a list of a data frame and a raster object returned from occurrence modules and covariate modules respectively. .data is passed automatically in workflow from the occurrence and covariate modules to the process module(s) and should not be passed by the user.

sd

standard deviation of the gaussian noise to add the the coordinates of each ocurrence record

Details

Currently only a single standard deviation is provided for all occurrence records. Future iterations of this module will enable more control over the uncertainty distributions

Data type

presence-only, presence/absence, presence/background, abundance, proportion

Version

0.1

Date submitted

2016-06-16

See also

Other process: AddRandomUniformPredictors, BackgroundAndCrossvalid, Background, Bootstrap, Clean, Crossvalidate, MESSMask, NoProcess, OneHundredBackground, OneThousandBackground, PartitionDisc, StandardiseCov, TargetGroupBackground, Transform, addInteraction

Author
Nick Golding, nick.golding.research@gmail.com