Occurrence module: LocalOccurrenceData

Occurrence module to format local occurrence data to be used with zoon. Must be a .csv file with three columns longitude, latitude and value in that order.

LocalOccurrenceData(filename = "myData.csv", occurrenceType = "presence",
  columns = c(long = "longitude", lat = "latitude", value = "value"),
  externalValidation = FALSE)

Arguments

filename

The path to the spreadsheet. The spreadsheet should have a header giving column names. The column names should be 'longitude', 'latitude', and 'value'. 'value' is a numeric such as a proportion, count or presence/absence (1/0). The file can be .csv, .tab, .tsv, or .xlsx

occurrenceType

What type data is it? One of 'presence', 'presence/absence', 'abundance', 'probability'

columns

Which columns in the spreadsheet relate to longitude, latitude and response value? Takes a named character vector e.g. c(long = 'longitude', lat = 'latitude', value = 'responseVals'). If an unnamed character vector is given, the order is assumed to be long, lat, value.

externalValidation

Logical indicating whether this data is external validation data. Cross validation is handled elsewhere and this argument should be left as FALSE for data that will be used in cross validation.

Version

1.0

Date submitted

2015-11-13

Data type

presence-only, presence/background, presence/absence, abundance, proportion

See also

Other occurrence: AnophelesPlumbeus, CWBZimbabwe, CarolinaWrenPA, CarolinaWrenPO, Lorem_ipsum_UK, NATrees, NBNdataByName, NaiveRandomPresenceAbsence, NaiveRandomPresence, SpOcc, SugarMaple, UKAnophelesPlumbeus

Author
ZOON Developers, zoonproject@gmail.com