Model module: MachineLearn

Model module to fit a very large number of machine learning models.

MachineLearn(.df, method = "glm", tuneLength = 8, metric = "ROC",
  number = 5, repeats = 1, tuneGrid = NULL, trControl = NULL, ...)

Arguments

.df

Internal parameter, do not use in the workflow function. .df is data frame that combines the occurrence

method

The machine learning method to use. Common examples are "glm", "nnet", "gbm", "glmnet", "rf". See http://topepo.github.io/caret/modelList.html for a full list. Only classification or dual use models are useable.

tuneLength

How many values of each tuning/hyperparameter should be tried?

metric

a string that specifies what summary metric will be used to select the optimal model. Options are "ROC", "Accuracy" and "Kappa".

number

How many folds to use in cross validation.

repeats

How many times should the entire cross validation process be repeated. Increasing this will reduce instability in your model performace, but will take longer to run.

tuneGrid

Explicitely pass a data frame of tuning/hyperparameter combinations. If NULL, tuneLength will be used instead.

trControl

A named list of further arguments to pass to trainControl. See trainControl for details.

...

Other arguments passed to train.

Version

1.0

Date submitted

2015-11-13

Data type

presence/absence

See also

caret::train trainControl

Other model: BiomodModel, GBM, LogisticRegression, MaxEnt, MaxLike, MyMaxLike, OptGRaF, QuickGRaF, RandomForest, StochasticLogisticRegression, mgcv

Author
Tim CD Lucas, timcdlucas@gmail.com