NBN Data by Common Name

This module supplies occurrence data from the UK NBN gateway. To get access to data you must first register at https://data.nbn.org.uk/User/Register. You will need your username and password when running this module. You can specify the data to retrieve by dataset, species, time, location and/or group.

NBNdataByName(species = "Myotis daubentonii", username = "ZOON",
  password = "zoonproject", datasets = NULL, startYear = 2000,
  endYear = 2000, VC = NULL, group = NULL, gridRef = NULL,
  acceptTandC = FALSE, silent = FALSE, attributes = FALSE)

Arguments

species

The latin name of a species to search for, or a vector of such names.

username

your NBN username

password

your NBN password

datasets

a list of dataset keys which are strings of 8 alphanumeric characters. Look up datasets here: https://data.nbn.org.uk/Datasets

startYear

a 4 digit integer year

endYear

a 4 digit integer year

VC

a string giving a vice-county name (see listVCs in the rnbn package)

group

a string giving the name of a group (see listGroups in the rnbn package). Using group will retireve data for all TVKs in this group. for example using group "reptile" will search using over 150 TVKs including TVKs for higher taxonomic groups such families within reptiles. Therefore it may be preferrable to search using a vector of species

gridRef

a string giving a UK gridreference in which to search for occurrences

acceptTandC

if set to TRUE you accept the NBN gateway terms and conditions and privacy policy. These can be found at https://data.nbn.org.uk/Terms. Accepting the terms and conditions supresses the corresponding warning message.

silent

if TRUE information returned to console is reduced

attributes

if FALSE then attribute data is not returned, this may improve the speed of large requests.

Details

For more detials see the rnbn package

Data type

presence-only

Version

0.1

Date submitted

2016-06-16

See also

Other occurrence: AnophelesPlumbeus, CWBZimbabwe, CarolinaWrenPA, CarolinaWrenPO, LocalOccurrenceData, Lorem_ipsum_UK, NATrees, NaiveRandomPresenceAbsence, NaiveRandomPresence, SpOcc, SugarMaple, UKAnophelesPlumbeus

Author
Tom August, tomaug@ceh.ac.uk