Covariate module: NCEP

Covariate module to grab coarse resolution environmental data from NCEP

NCEP(extent = c(-5, 5, 50, 60), variables = "hgt", status.bar = FALSE)

Arguments

extent

A numeric vector of length 4 giving the coordinates of the rectangular region within which to carry out the analysis, in the order: xmin, xmax, ymin, ymax.

variables

Character vector of which variables to select. See NCEP.gather for more options but basic options are *'air'* Air Temperature deg K *'hgt'* Geopotential Height m *'rhum'* Relative Humidity *'shum'* Specific Humidity kg/kg *'omega'* Omega [Vertical Velocity] Pascal/s *'uwnd'* U-Wind Component [East/West] m/s *'vwnd'* V-Wind Component [North/South] m/s

status.bar

Logical. Should a status bar be shown indicating the percentage of completion?

Version

1.0

Date submitted

2015-11-13

See also

RNCEP::NCEP.gather

Other covariate: AirNCEP, Bioclim_future, Bioclim, CarolinaWrenRasters, LocalRaster, NaiveRandomRaster, UKAir, UKBioclim

Author
ZOON Developers, zoonproject@gmail.com