an ROC-curve with the AUC value plotted in the title

The module retrieves output from a presence only, or presence-absence model and outputs an ROC-curve with the AUC-value above it it uses the evaluate function from dismo.

ROCcurve(.model, .ras, newwin = TRUE)

Arguments

.model

Internal parameter, do not use in the workflow function. .model is list of a data frame (data) and a model object (model). .model is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.

.ras

Internal parameter, do not use in the workflow function. .ras is a raster layer, brick or stack object. .ras is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.

newwin

logical indicator (default: true) to specify whether the plot should be made in a new window

Details

The module writes the ROC-plot to whatever is opened by dev.new() this can be changed to plot in the current output device by setting the option newwin=FALSE

Data type

presence-only, presence/absence, presence/background

Version

0

Date submitted

2016-06-15

See also

Other output: Appify, InteractiveCovariateMap, InteractiveMap, NoOutput, PerformanceMeasures, PredictNewRasterMap, PrintMap, ResponseCurveViz, ResponseCurve, SameTimePlaceMap, SeparatePA, SurfaceMap, VariableImportance

Author
E.E. van Loon, e.e.vanloonD@uva.nl