Output module: ResponseCurve

Plot a detailed conditional response curve from the model against one covariate.

ResponseCurve(.model, .ras, cov = NULL)

Arguments

.model

Internal parameter, do not use in the workflow function. .model is list of a data frame (data) and a model object (model). .model is passed automatically in workflow, combining data from the model module(s) and process module(s), to the output module(s) and should not be passed by the user.

.ras

Internal parameter, do not use in the workflow function. .ras is a raster layer, brick or stack object. .ras is passed automatically in workflow from the covariate module(s) to the output module(s) and should not be passed by the user.

cov

which of the covariates to plot the response against. If NULL, the function is executed for all covariates, one after another. Else it should be a single numeric identifying the covariate to plot.

Version

1.0

Date submitted

2015-11-13

See also

Other output: Appify, InteractiveCovariateMap, InteractiveMap, NoOutput, PerformanceMeasures, PredictNewRasterMap, PrintMap, ROCcurve, ResponseCurveViz, SameTimePlaceMap, SeparatePA, SurfaceMap, VariableImportance

Author
ZOON Developers, zoonproject@gmail.com