Occurrence module: SpOcc

Occurrence module to collect occurrence data from a number of data bases.

SpOcc(species = "Myotis daubentonii", extent = c(-1, 0, 51, 52),
  databases = "gbif", type = "presence", limit = 10000)

Arguments

species

A character string giving the species name. e.g. 'Anopheles plumbeus'.

extent

A numeric vector of length 4 giving the coordinates of the rectangular region within which to carry out the analysis, in the order: xmin, xmax, ymin, ymax.

databases

A character vector giving the databases to use. Choose from "gbif", "bison", "inat", "ebird", "ecoengine", "antweb". Defaults to gbif. NB I have had some problems with databases other than gbif.

type

A string giving the type of occurrence data to treat these records as. Must be one of 'presence' (the default and most common use case), 'absence' or 'background' (which may be used to implement a 'target background group' approach for overcoming observation bias).

limit

The maximum number of occurrence records to return.

Version

1.0

Date submitted

2015-11-13

Data type

presence-only, presence/absence, abundance, proportion

See also

spocc::occ

Other occurrence: AnophelesPlumbeus, CWBZimbabwe, CarolinaWrenPA, CarolinaWrenPO, LocalOccurrenceData, Lorem_ipsum_UK, NATrees, NBNdataByName, NaiveRandomPresenceAbsence, NaiveRandomPresence, SugarMaple, UKAnophelesPlumbeus

Author
ZOON Developers, zoonproject@gmail.com