StochasticLogisticRegression

Fits a simple logistic regression model, but then samples a single draw of the model coefficients (with probability proportional to their likelihood) and uses these for future prediction

StochasticLogisticRegression(.df)

Arguments

.df

Internal parameter, do not use in the workflow function. .df is data frame that combines the occurrence data and covariate data. .df is passed automatically in workflow from the process module(s) to the model module(s) and should not be passed by the user.

Details

Coefficients are simulated from the likelihood density using the approximate hessian matrix, under an assumption of multivariate normality. This module is intended for use in a Monte Carlo simulation procedure to propagate uncertainty through an analysis. It is not intended to be used on its own!

Data type

presence/absence

Version

0.1

Date submitted

2016-06-16

See also

Other model: BiomodModel, GBM, LogisticRegression, MachineLearn, MaxEnt, MaxLike, MyMaxLike, OptGRaF, QuickGRaF, RandomForest, mgcv

Author
Nick Golding, nick.golding.research@gmail.com