Sugar Maple distribution in North America

Presence/absence of Sugar Maple (Acer saccharum) in North America. Each record corresponds to the centroid of a 0.5 degree cell.

SugarMaple()

Details

The extent of the occurrence data is c(-170, -20, 10, 80). These data were used as part of a wider analysis by Morin & Thuiller (2009), and were subsequently analysed and made available by Talluto et al.(2016). Morin, X., & Thuiller, W. (2009). Comparing niche-and process-based models to reduce prediction uncertainty in species range shifts under climate change. Ecology, 90(5), 1301–1313. http://doi.org/10.1890/08-0134.1 Talluto, M. V., Boulangeat, I., Ameztegui, A., Aubin, I., Berteaux, D., Butler, A., … Gravel, D. (2016). Cross-scale integration of knowledge for predicting species ranges: A metamodelling framework. Global Ecology and Biogeography, 25(2), 238–249. http://doi.org/10.1111/geb.12395

Data type

presence/absence

Version

0.2

Date submitted

2016-06-15

See also

Other occurrence: AnophelesPlumbeus, CWBZimbabwe, CarolinaWrenPA, CarolinaWrenPO, LocalOccurrenceData, Lorem_ipsum_UK, NATrees, NBNdataByName, NaiveRandomPresenceAbsence, NaiveRandomPresence, SpOcc, UKAnophelesPlumbeus

Author
Matt Talluto & Nick Golding, mtalluto@gmail.com